+++ src/umpire/resource/MemoryResourceTypes.hpp
@@ -23,7 +23,7 @@ struct MemoryResourceTypeHash {
   }
 };
 
-enum MemoryResourceType { Host, Device, Unified, Pinned, Constant, File, Unknown };
+enum MemoryResourceType { Host, Device, Unified, Pinned, Constant, File, Shared, Unknown };
 
 inline std::string resource_to_string(MemoryResourceType type)
 {
@@ -40,6 +40,8 @@ inline std::string resource_to_string(MemoryResourceType type)
       return "DEVICE_CONST";
     case File:
       return "FILE";
+    case Shared:
+      return "SHARED";
     default: 
       UMPIRE_ERROR("Unkown resource type: " << type);
       return "UNKNOWN"; 
@@ -54,6 +56,7 @@ inline MemoryResourceType string_to_resource(const std::string& resource)
   else if (resource == "PINNED") return MemoryResourceType::Pinned;
   else if (resource == "DEVICE_CONST") return MemoryResourceType::Constant;
   else if (resource == "FILE") return MemoryResourceType::File;
+  else if (resource == "SHARED") return MemoryResourceType::Shared;
   else {
     UMPIRE_ERROR("Unkown resource name: " << resource);
     return MemoryResourceType::Unknown;